climbing: extract consts
