+++ src/components/Map/climbingTiles/consts.ts
@@ -1,0 +1,31 @@
+export const AREA = {
+  HAS_IMAGES: {
+    IMAGE: 'climbing:area-blue',
+    COLOR: 'rgba(0, 59, 210, 1)',
+  },
+  NO_IMAGES: {
+    IMAGE: 'climbing:area-gray',
+    COLOR: '#666',
+  },
+};
+
+export const CRAG = {
+  HAS_IMAGES: {
+    IMAGE: 'climbing:crag-red',
+    COLOR: '#ea5540',
+  },
+  NO_IMAGES: {
+    IMAGE: 'climbing:crag-gray',
+    COLOR: '#666',
+  },
+};
+
+export const VIA_FERRATA = {
+  IMAGE: 'climbing:via-ferrata-gray',
+  COLOR: '#666',
+};
+
+export const GYM = {
+  IMAGE: 'climbing:climbing-gym-gray',
+  COLOR: '#666',
+};
